requirements.txt
pyyaml
click
pytest
